Move displayNotifications function into helper file
